+++ webpack/published-js-components.js
@@ -1,18 +1,14 @@
 const publishedJSComponents = {
-  VirtualizedList: "components/VirtualizedList/VirtualizedList.jsx",
-  VirtualizedGrid: "components/VirtualizedGrid/VirtualizedGrid.jsx",
   ColorPicker: "components/ColorPicker/ColorPicker.jsx",
   ColorPickerContent: "components/ColorPicker/components/ColorPickerContent/ColorPickerContentComponent.jsx",
   Tipseen: "components/Tipseen/Tipseen.jsx",
   TipseenContent: "components/Tipseen/TipseenContent.jsx",
   TipseenWizard: "components/Tipseen/TipseenWizard.jsx",
   TipseenImage: "components/Tipseen/TipseenImage.jsx",
-  Steps: "components/Steps/Steps.jsx",
   Combobox: "components/Combobox/Combobox.jsx",
   ResponsiveList: "components/ResponsiveList/ResponsiveList.jsx",
   EditableHeading: "components/EditableHeading/EditableHeading.jsx",
   EditableInput: "components/EditableInput/EditableInput.jsx",
-  ExpandCollapse: "components/ExpandCollapse/ExpandCollapse.jsx",
   MultiStepIndicator: "components/MultiStepIndicator/MultiStepIndicator.jsx",
   MenuTitle: "components/Menu/MenuTitle/MenuTitle.jsx",
   MenuItem: "components/Menu/MenuItem/MenuItem.jsx",
@@ -22,20 +18,11 @@ const publishedJSComponents = {
   MenuGridItem: "components/Menu/MenuGridItem/MenuGridItem.jsx",
   Dialog: "components/Dialog/Dialog.jsx",
   MenuButton: "components/MenuButton/MenuButton.jsx",
-  SplitButton: "components/SplitButton/SplitButton.jsx",
-  Checkbox: "components/Checkbox/Checkbox.js",
+  Checkbox: "components/Checkbox/Checkbox.jsx",
   Dropdown: "components/Dropdown/Dropdown.jsx",
-  Toast: "components/Toast/Toast.jsx",
   Flex: "components/Flex/Flex.jsx",
-  LinearProgressBar: "components/ProgressBars/LinearProgressBar/LinearProgressBar.jsx",
   GridKeyboardNavigationContext: "components/GridKeyboardNavigationContext/GridKeyboardNavigationContext.js",
-  allIcons: "components/Icon/Icons",
-  Skeleton: "components/Skeleton/Skeleton.jsx",
-  Slider: "components/Slider/Slider.jsx",
-  Modal: "components/Modal/Modal.jsx",
-  ModalHeader: "components/ModalHeader/ModalHeader.jsx",
-  ModalFooter: "components/ModalFooter/ModalFooter.jsx",
-  ModalContent: "components/ModalContent/ModalContent.jsx"
+  allIcons: "components/Icon/Icons"
 };
 // eslint-disable-next-line no-undef
 module.exports = {